+++ src/main/java/ru/alishev/springcourse/models/User.java
@@ -14,7 +14,7 @@ public class User implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
     private String name;
     private String password;
     @Transient
@@ -29,18 +29,18 @@ public class User implements UserDetails {
 
     }
 
-    public User(int id, String name, String password, Set<Role> roles) {
+    public User(Long id, String name, String password, Set<Role> roles) {
         this.id = id;
         this.name = name;
         this.password = password;
         this.roles = roles;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
